Medha

+++ src/app/components/inventory/sub-cat-listing/sub-cat-listing.component.ts
@@ -7,6 +7,7 @@ import { MatSnackBar, } from '@angular/material';
 import { AuthService } from '../../../core/auth.service';
 import { AddNewComponent } from '../../common/add-new/add-new.component';
 import {MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
+import { Message } from '@angular/compiler/src/i18n/i18n_ast';
 
 @Component({
   selector: 'app-sub-cat-listing',
@@ -20,7 +21,7 @@ export class SubCatListingComponent implements OnInit {
   nColumns: number;
   items: Observable<Item[]>;
 
-  constructor(public snackBar: MatSnackBar,private dialog: MatDialog, private auth: AuthService,private dataService: DataService, private router: Router) {
+  constructor(public snackBar: MatSnackBar, private dialog: MatDialog, private auth: AuthService, private dataService: DataService, private router: Router) {
     const mWidth = window.innerWidth;
     this.setWidth(mWidth);
   }
@@ -66,38 +67,37 @@ export class SubCatListingComponent implements OnInit {
     this.router.navigate(['item-details', id]);
   }
 
-  popUp(message: string,action: string) {
-    this.snackBar.open(message,action,{
-      duration:2500,
+  popUp(message: string, action: string) {
+    this.snackBar.open(message, action, {
+      duration: 2500,
     });
   }
 
- 
-    
   onDelete(id) {
-    this.auth.user.take(1).subscribe(val => {
-      if (this.auth.canDelete(val)) {
-    this.dataService.deleteItemById(id).then(() => {
-      console.log('deleted item succesfully');
-    }).catch(err => {
-      console.error('error while deletng', err);
-      alert('error in deleting');
-    });
-  }
-  else{
-    this.popUp('Not Admin : ','No Access to Delete');
-  }
-});
+    const p = prompt('If you want to delete the item enter Y ');
+    console.log(p);
+    if ( p === 'Y' || p === 'y') {
+      this.auth.user.take(1).subscribe(val => {
+        if (this.auth.canDelete(val)) {
+      this.dataService.deleteItemById(id).then(() => {
+        console.log('deleted item succesfully');
+      }).catch(err => {
+        console.error('error while deletng', err);
+        alert('error in deleting');
+      });
+    } else {
+      this.popUp('Not Admin : ', 'No Access to Delete');
+    }});
+    }
   }
 
   onEdit(itm) {
-    console.log('item is',itm);
+    console.log('item is', itm);
     this.auth.user.take(1).subscribe(val => {
       if (this.auth.canEdit(val)) {
         const dialogRef = this.dialog.open(AddNewComponent, {
           width: '450px',
           data: {
-           
             'item': itm
           },
           disableClose: false
